feat: 로그인 Service 구현

+++ server/src/main/java/com/taxisharing/server/auth/dto/BasicUserInfo.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class BasicUserInfo {
-    private Long id;
+    private Integer id;
     private String nickname;
 
-    public BasicUserInfo(Long id, String nickname)
+    public BasicUserInfo(Integer id, String nickname)
     {
         this.id = id;
         this.nickname = nickname;
